Build v0.0.1 	deleted:    .circleci/CompileKernel.sh 	modified:   .circleci/build.sh 	modified:   .circleci/setup.sh 	deleted:    .github/workflows/test.yml 	modified:   arch/arm64/configs/lisa_defconfig 	deleted:    test.sh

+++ .circleci/build.sh
@@ -2,9 +2,6 @@
 #
 # Compile script for QuicksilveR kernel
 # Copyright (C) 2020-2021 Adithya R.
-# (edits for CrystalCore kernel @dkpost3)
-
-export PATH="$TC_DIR"/bin:"$PATH"
 
 ##----------------------------------------------------------##
 
@@ -30,93 +27,53 @@ tg_post_build()
 	-F caption="$2 | *MD5 Checksum : *\`$MD5CHECK\`"
 }
 
-##----------------------------------------------------------##
-
 MODEL="Xiaomi 11 Lite 5G NE"
 DEVICE="lisa"
 ARCH=arm64
 BOT_MSG_URL="https://api.telegram.org/bot$token/sendMessage"
 BOT_BUILD_URL="https://api.telegram.org/bot$token/sendDocument"
-SECONDS=0 # builtin bash timer
-PROCS=$(nproc --all)
 CI="Circle CI"
-CHATID="-1001293242785"
-DEFCONFIG="lisa_defconfig"
-MAKE_PARAMS1="ARCH=arm64 CC=$TC_DIR/bin/clang CLANG_TRIPLE=$TC_DIR/bin/aarch64-linux-gnu- LD=$TC_DIR/bin/ld.lld LLVM=1 LLVM_IAS=1 \
-	CROSS_COMPILE=$TC_DIR/bin/llvm-"
-
-# Set a commit head
 COMMIT_HEAD=$(git log --oneline -1)
-
-#Check Kernel Version
 KV=$(make $MAKE_PARAMS1 kernelversion)
+KBUILD_COMPILER_STRING=$("$TC_DIR"/bin/clang --version | head -n 1 | perl -pe 's/\(http.*?\)//gs' | sed -e 's/  */ /g' -e 's/[[:space:]]*$//')
+export PATH="$TC_DIR/bin:$PATH"
 
-##------------------------------------------------------##
+##----------------------------------------------------------##
 
-exports()
-{
+SECONDS=0 # builtin bash timer
+TC_DIR="$BASE_DIR/clang"
+AK3_DIR="$BASE_DIR/AnyKernel3"
+DEFCONFIG="lisa_defconfig"
+output="$BASE_DIR/Kernel/out"
 
-       KBUILD_COMPILER_STRING=$("$TC_DIR"/bin/clang --version | head -n 1 | perl -pe 's/\(http.*?\)//gs' | sed -e 's/  */ /g' -e 's/[[:space:]]*$//')
-		   PATH=$TC_DIR/bin/:$PATH
-
-export KBUILD_BUILD_USER ARCH SUBARCH PATH \
-		   KBUILD_COMPILER_STRING BOT_MSG_URL \
-		   BOT_BUILD_URL PROCS
-
-# shellcheck source=/etc/os-release
-export DISTRO=$(source /etc/os-release && echo "${NAME}")
-export KBUILD_BUILD_HOST=$(uname -a | awk '{print $2}')
-export CI_BRANCH=$(git rev-parse --abbrev-ref HEAD)
-TERM=xterm
-
-## Check for CI
-if [ "$CI" ]
-then
-	if [ "$CIRCLECI" ]
-	then
-		export KBUILD_BUILD_VERSION=$CIRCLE_BUILD_NUM
-		export KBUILD_BUILD_HOST="CircleCI"
-		export CI_BRANCH=$CIRCLE_BRANCH
-	fi
-	if [ "$DRONE" ]
-	then
-		export KBUILD_BUILD_VERSION=$DRONE_BUILD_NUMBER
-		export KBUILD_BUILD_HOST=$DRONE_SYSTEM_HOST
-		export CI_BRANCH=$DRONE_BRANCH
-		export BASEDIR=$DRONE_REPO_NAME
-		export SERVER_URL="${DRONE_SYSTEM_PROTO}://${DRONE_SYSTEM_HOSTNAME}/${AUTHOR}/${BASEDIR}/${KBUILD_BUILD_VERSION}"
-	else
-		tg_post_msg "<b>##----------------------------------------------------------##</b>"
-	fi
-fi
-
-BLDV="R0.3-v0.0.0"
+BLDV="R0.3-v0.0.1"
 ZIPNAME="Proton-$BLDV.zip"
-ZIPSIGNED="Proton-$BLDV-signed.zip"
 
-MAKE_PARAMS="O=out ARCH=arm64 CC=$TC_DIR/bin/clang CLANG_TRIPLE=$TC_DIR/bin/aarch64-linux-gnu- LD=$TC_DIR/bin/ld.lld LLVM=1 LLVM_IAS=1 \
+MAKE_PARAMS="O=out ARCH=arm64 CC=clang CLANG_TRIPLE=aarch64-linux-gnu- LLVM=1 LLVM_IAS=1 \
 	CROSS_COMPILE=$TC_DIR/bin/llvm-"
 
-make $MAKE_PARAMS mrproper
-make $MAKE_PARAMS $DEFCONFIG
-cp "$OUTPUT"/.config $KERNEL_SRC/arch/arm64/configs/lisa_defconfig
-tg_post_build "$KERNEL_SRC"/.config
-tg_post_msg "<b>Successfully regenerated defconfig at $DEFCONFIG</b>"
+MAKE_PARAMS1="ARCH=arm64 CC=clang CLANG_TRIPLE=aarch64-linux-gnu- LLVM=1 LLVM_IAS=1 \
+	CROSS_COMPILE=$TC_DIR/bin/llvm-"
 
+if [[ $1 = "-r" || $1 = "--regen" ]]; then
+	make $MAKE_PARAMS $DEFCONFIG
+	cp .config arch/arm64/configs/$DEFCONFIG
+	tg_post_build .config
+	tg_post_build out/.config
+	echo -e "\nSuccessfully regenerated defconfig at $DEFCONFIG"
+	exit
+fi
 
 if [[ $1 = "-c" || $1 = "--clean" ]]; then
 	rm -rf out
 	echo "Cleaned output folder"
 fi
 
-# New defconfig to use instead
-LISADEF="lisa_defconfig"
-
 mkdir -p out
 make $MAKE_PARAMS $DEFCONFIG
 
-tg_post_msg "<b>$KBUILD_BUILD_VERSION CI Build Triggered</b>%0A<b>Docker OS: </b><code>$DISTRO</code>%0A<b>Kernel Version : </b><code>$KV</code>%0A<b>Date : </b><code>$(TZ=Asia/Jakarta date)</code>%0A<b>Device : </b><code>$MODEL [$DEVICE]</code>%0A<b>Pipeline Host : </b><code>$CI</code>%0A<b>Host Core Count : </b><code>$PROCS</code>%0A<b>Compiler Used : </b><code>$KBUILD_COMPILER_STRING</code>%0A<b>Linker : </b><code>$LINKER</code>%0a<b>Branch : </b><code>$CI_BRANCH</code>%0A<b>Top Commit : </b><code>$COMMIT_HEAD</code>%0A<a href='$SERVER_URL'>Link</a>"
 tg_post_msg "<b>Starting compilation</b>"
+tg_post_msg "<b>$KBUILD_BUILD_VERSION CI Build Triggered</b>%0A<b>Docker OS: </b><code>$DISTRO</code>%0A<b>Kernel Version : </b><code>$KV</code>%0A<b>Date : </b><code>$(TZ=Asia/Jakarta date)</code>%0A<b>Device : </b><code>$MODEL [$DEVICE]</code>%0A<b>Pipeline Host : </b><code>$CI</code>%0A<b>Host Core Count : </b><code>$PROCS</code>%0A<b>Compiler Used : </b><code>$KBUILD_COMPILER_STRING</code>%0A<b>Linker : </b><code>$LINKER</code>%0a<b>Branch : </b><code>$CI_BRANCH</code>%0A<b>Top Commit : </b><code>$COMMIT_HEAD</code>%0A<a href='$SERVER_URL'>Link</a>"
 make -j$(nproc --all) $MAKE_PARAMS || exit $?
 make -j$(nproc --all) $MAKE_PARAMS INSTALL_MOD_PATH=modules INSTALL_MOD_STRIP=1 modules_install
 
@@ -125,24 +82,29 @@ dtb="out/arch/arm64/boot/dts/vendor/qcom/yupik.dtb"
 dtbo="out/arch/arm64/boot/dts/vendor/qcom/lisa-sm7325-overlay.dtbo"
 
 if [ -f "$kernel" ] && [ -f "$dtb" ] && [ -f "$dtbo" ]; then
-	tg_post_msg "<b>Kernel compiled succesfully!</b>"
-fi
-	cp "$kernel" "$AK3_DIR"/Image
-	cp "$dtb" "$AK3_DIR"/dtb
-	tg_post_msg "<b>Creating DTBO Image</b>"
-	python3 scripts/dtc/libfdt/mkdtboimg.py create "$AK3_DIR"/dtbo.img --page_size=4096 $dtbo
-	cp "$(find $OUTPUT/modules/lib/modules/5.4* -name '*.ko')" "$AK3_DIR"/modules/vendor/lib/modules/
-	cp $OUTPUT/modules/lib/modules/5.4*/modules.{alias,dep,softdep} $AK3_DIR/modules/vendor/lib/modules
-	cp $OUTPUT/modules/lib/modules/5.4*/modules.order $AK3_DIR/modules/vendor/lib/modules/modules.load
-	sed -i 's/\(kernel\/[^: ]*\/\)\([^: ]*\.ko\)/\/vendor\/lib\/modules\/\2/g' "$AK3_DIR"/modules/vendor/lib/modules/modules.dep
-	sed -i 's/.*\///g' "$AK3_DIR"/modules/vendor/lib/modules/modules.load
-	rm -rf out/arch/arm64/boot out/modules
-	tg_post_msg "<b>!Zipping Up!</b>"
-	cd "$AK3_DIR"
-	zip -r9 "$ZIPNAME" * -x ".git" -x ".github" -x "README.md" -x "*placeholder"
-	echo "Zip: "$ZIPNAME""
+	echo -e "\nKernel compiled succesfully! Zipping up...\n"
+	if [ -d "$AK3_DIR" ]; then
+		cp -r $AK3_DIR AnyKernel3
+		git -C AnyKernel3 checkout lisa &> /dev/null
+	elif ! git clone -q https://github.com/ghostrider-reborn/AnyKernel3 -b lisa; then
+		echo -e "\nAnyKernel3 repo not found locally and couldn't clone from GitHub! Aborting..."
+		exit 1
+	fi
+cp $kernel AnyKernel3
+cp $dtb AnyKernel3/dtb
+python3 scripts/dtc/libfdt/mkdtboimg.py create AnyKernel3/dtbo.img --page_size=4096 $dtbo
+cp $(find out/modules/lib/modules/5.4* -name '*.ko') AnyKernel3/modules/vendor/lib/modules/
+cp out/modules/lib/modules/5.4*/modules.{alias,dep,softdep} AnyKernel3/modules/vendor/lib/modules
+cp out/modules/lib/modules/5.4*/modules.order AnyKernel3/modules/vendor/lib/modules/modules.load
+sed -i 's/\(kernel\/[^: ]*\/\)\([^: ]*\.ko\)/\/vendor\/lib\/modules\/\2/g' AnyKernel3/modules/vendor/lib/modules/modules.dep
+sed -i 's/.*\///g' AnyKernel3/modules/vendor/lib/modules/modules.load
+rm -rf out/arch/arm64/boot out/modules
+	cd AnyKernel3
+	zip -r9 "$ZIPNAME" * -x .git README.md *placeholder
+	echo "Zip: $ZIPNAME"
 	tg_post_build "$ZIPNAME"
-	cp "$ZIPNAME" "$UPLOADFOLDER"
-	tg_post_msg "<b>! "$ZIMPNAME" Completed in $((SECONDS / 60)) minute(s) and $((SECONDS % 60)) second(s) !</b>"
-	cd "$UPLOADFOLDER"
-	curl https://raw.githubusercontent.com/KazuDante89/Cirrus-CI/main/upload.sh | python3 -
+	cd ..
+	rm -rf AnyKernel3
+	echo -e "\nCompleted in $((SECONDS / 60)) minute(s) and $((SECONDS % 60)) second(s) !"
+	exit 1
+fi
